Update SubBoom.cs
Made a new class for DepthCharge. Stuck on grabbing the index of depthCharges List and applying the index to the GameObject currentDepthCharge so that we are able to move it downwards linearly.

Let me know if this is the right place to update code for the future. Still learning my way around GitHub!