Added Explosion Effect Class
I changed the ExplodeCharge() to make it it's own class called ExplosionEffect since we would've had to make a construction class for it anyways. Right now, it just makes the outline of the explosion as it has problems with using 2 sprites at the same time, so we could either just make a separate sprite for just the explosion which is the square outline with the orange circle, or find another way to code those two sprites in as separate things.