added player to scene, player's depth charge into scene, depth charge particles,
added the player (destroyer) in the scene along with the depth charges using the space key like before. i also added the particle system used on the submarine with the depth charge. right now though, the depth charge cannot explode at the right time so that will need to be done.