created torpedo pooler and  worked on torpedo controller
i forgot to publish the rest of depth charge movement changes to main, so i'll just push this last thing, which is unrelated, into this branch and then i'll push this whole branch to main.

i added a way to pool torpedos for the submarines, but right now they're stuck being hidden when the game is run. i also added some stuff to the torpedo controller so that it can move and, once we get a sound for it, will play a sound whenever a submarine launches a torpedo (something quick but not annoying since it will be repeated a lot in game), and i updated enemy controller to add torpedos to it.