deleted unessecary child classes and lowered explosion sound volume
i realized we didn't need most of the child classes since it's not like we need to override any methods or anything, so the only child class i kept was the submarine one. i also lowered the volume of the explosion sound effects, but it sounds like it didn't change anything, but i'll figure out a way to lower it.