points score is being added properly, player can now see how far the depth charge travels by holding the space bar, and optimizing game data
I got the points to add properly when the player successfully explodes a submarine! In addition, I added the depth charge today that is a child to the player object, and will show up when the player holds down the space bar to see how far the "real" depth charge will travel once they release the space bar. And finally, I'm currently optimizing the game data script right now, but it has some bugs since I'm not finished with it yet, so I made a branch just in case it's not working out.