Added all sounds in the game, Added AudioManager and VolumeManager, cleaned up some code, and added a Pause feature
All sounds and music are in the game and are accessed through the AudioManager script file! the audio can also be adjusted in the main menu and in-game with the VolumeManager script, whether by going in the settings tab in the main menu or with the 'Escape' key in-game. Hitting the 'Escape' key also pauses the game!

However, I encountered some audio bugs while implementing everything, and a lot of the audio clips will have to be balanced (some are too quiet, some are too loud) and the explosion sound effect sounds really weird in-game for some reason.