fixed some audio bugs and made the settings a prefab
Adjusted all audio to be balanced, though it probably still needs some tweaking. In addition, some of the audio clips are still sounding wrong for some reason (most notably the explosion sfx and game over sfx)

Fixed an audio bug where the game would crash when the player would return to the settings in the menu after losing.

Added the audio save data to the Game Data script instead of PlayerPrefs.

Currently, there is a bug where the sound settings arent saving when the player restarts (something about the savedata in the Game Data file) in addition to not all explosion sound effects playing when they're supposed to.